refactor: extract board fleet module
